// projects/ui/src/lib/components/po-accordion/interfaces/po-accordion-literals.interface.ts
/**
 * @usedBy PoAccordionComponent
 *
 * @description
 *
 * Interface para definição das literais usadas no `po-accordion`.
 */
export interface PoAccordionLiterals {
  /** Label do gerenciador de Accordion para expandir todos os itens. */
  expandAllItems?: string;

  /** Label do gerenciador de Accordion para colapsar todos os itens */
  closeAllItems?: string;
}